isolate block selector from blocks drawer and export
